feat: Add MinIO S3 support and improve Docker test environment
- Configure S3/MinIO integration with PXF servers
- Add Parquet profile support in pxf-profiles.xml
- Update cloud access test cases to use MinIO protocol
- Configure Gradle and Go to use aliyun mirrors for faster builds
- Improve Docker entrypoint with MinIO deployment
- Update test utilities to support protocol switching (s3/minio/hdfs)
- Fix OutOfMemory test server configuration

+++ automation/src/test/java/org/greenplum/pxf/automation/features/BaseWritableFeature.java
@@ -18,7 +18,9 @@ public class BaseWritableFeature extends BaseFeature {
     @Override
     protected void beforeClass() throws Exception {
         super.beforeClass();
-        hdfsWritePath = hdfs.getWorkingDirectory() + "/writable_results/";
+        if (hdfs != null) {
+            hdfsWritePath = hdfs.getWorkingDirectory() + "/writable_results/";
+        }
     }
 
     /**
@@ -27,6 +29,8 @@ public class BaseWritableFeature extends BaseFeature {
     @Override
     protected void afterMethod() throws Exception {
         super.afterMethod();
-        hdfs.removeDirectory(hdfsWritePath);
+        if (hdfs != null && hdfsWritePath != null) {
+            hdfs.removeDirectory(hdfsWritePath);
+        }
     }
 }
